Feat/tissue (#23)
* fix(dtype): standardize crops.zarr background to u16, drop f64

- mupattern-rs: background stored/read as u16; remove create_array_f64, store_chunk_f64
- mupattern-py: background, expression, tissue all uint16
- expression: keep background as u16/int in CSV (no f64 cast)


* tissue: read background as int from uint16 zarr


* feat(tissue): CellSAM integration, tissue tab and task in desktop

- mupattern-rs: add cellsam-rs dep, --method cellpose|cellsam, median fallback
- mupattern-py: drop _segment_frame_watershed, explicit cellpose/cellsam check
- mupattern-desktop: Tissue tab (GFP+ chart), Tissue task config modal, IPC


---------

+++ mupattern-py/src/mupattern_py/apps/expression/core.py
@@ -26,7 +26,7 @@ def run_analyze(
 
     bg_arr = root[f"pos/{pos:03d}/background"]
 
-    rows: list[tuple[int, str, int, int, float]] = []
+    rows: list[tuple[int, str, int, int, int]] = []
     total = len(crop_ids)
     for i, crop_id in enumerate(crop_ids):
         arr = crop_grp[crop_id]
@@ -35,7 +35,7 @@ def run_analyze(
 
         for t in range(n_times):
             intensity = int(np.array(arr[t, channel, 0]).sum())
-            background = float(bg_arr[t, channel, 0])  # per-pixel
+            background = int(bg_arr[t, channel, 0])  # per-pixel (crops.zarr background is uint16)
             rows.append((t, crop_id, intensity, area, background))
 
         if on_progress and total > 0:
